+++ src/algorithm/array/slidingWindow/__test__/checkInclusion.test.ts
@@ -1,0 +1,12 @@
+import { checkInclusion } from '../checkInclusion';
+
+test('checkInclusion', () => {
+  expect(checkInclusion('eidbaooo', 'abo'))
+    .toBeTruthy();
+
+  expect(checkInclusion('eidbaooo', 'ba'))
+    .toBeTruthy();
+
+  expect(checkInclusion('eicidbaooo', 'id'))
+    .toBeFalsy();
+});